requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
billiard==4.2.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.7.1
cryptography==44.0.2
debugpy==1.8.13
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.7
django-cachalot==2.7.0
django-cacheops==7.1
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-filter==25.1
django-jet-reboot==1.3.10
django-object-actions==4.3.0
django-redis==5.4.0
django-rest-passwordreset==1.5.0
django-silk==5.3.2
django-sslserver==0.22
django-sslserver-v2==1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docker==7.1.0
dotenv==0.9.9
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
filelock==3.18.0
freezegun==1.5.1
funcy==2.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gprof2dot==2024.6.6
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
identify==2.6.9
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.1
MarkupSafe==3.0.2
mongoengine==0.29.1
mypy-extensions==1.0.0
ngrok==1.4.0
nodeenv==1.9.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pymongo==4.11.3
pyparsing==3.2.3
pytest==8.3.5
pytest-celery==1.2.0
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-mock==3.14.0
pytest_docker_tools==3.1.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
sentry-sdk==2.24.1
setuptools==78.0.2
six==1.17.0
social-auth-app-django==5.4.3
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.5.6
social-auth-storage-mongoengine==1.0.1
sqlparse==0.5.3
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.29.3
vk==3.0
vk-api==11.9.9
wcwidth==0.2.13
Werkzeug==3.1.3
yamllint==1.37.0
